# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface ShopifyProductImage {
  url: string;
  /** @nullable */
  altText: string | null;
}

export interface ShopifyProductVariant {
  id: string;
  title: string;
  price: number;
  currencyCode: string;
  availableForSale: boolean;
}

export interface ShopifyProduct {
  id: string;
  title: string;
  handle: string;
  description: string;
  image: ShopifyProductImage | null;
  price: number;
  currencyCode: string;
  variants: ShopifyProductVariant[];
}

export interface ShopifyCartLineInput {
  merchandiseId: string;
  /**
     * @minimum 1
     * @maximum 99
     */
  quantity: number;
}

export interface ShopifyCartInput {
  /** @minItems 1 */
  lines: ShopifyCartLineInput[];
}

export interface ShopifyCartLinesInput {
  /** @minItems 1 */
  lines: ShopifyCartLineInput[];
}

export interface ShopifyCartLineUpdate {
  /**
     * @minimum 0
     * @maximum 99
     */
  quantity: number;
}

export interface ShopifyCartLine {
  id: string;
  quantity: number;
  merchandiseId: string;
  title: string;
  variantTitle: string;
  price: number;
  currencyCode: string;
  image: ShopifyProductImage | null;
}

export interface ShopifyCart {
  id: string;
  checkoutUrl: string;
  lines: ShopifyCartLine[];
  subtotal: number;
  currencyCode: string;
}

export type ListShopifyProductsParams = {
/**
 * @minimum 1
 * @maximum 50
 */
first?: number;
};