# lib/api-zod/src/generated/types/shopifyCart.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ShopifyCartLine } from './shopifyCartLine';

export interface ShopifyCart {
  id: string;
  checkoutUrl: string;
  lines: ShopifyCartLine[];
  subtotal: number;
  currencyCode: string;
}