# lib/api-zod/src/generated/types/shopifyCartInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ShopifyCartLineInput } from './shopifyCartLineInput';

export interface ShopifyCartInput {
  /** @minItems 1 */
  lines: ShopifyCartLineInput[];
}