# lib/api-zod/src/generated/types/shopifyCartLine.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ShopifyProductImage } from './shopifyProductImage';

export interface ShopifyCartLine {
  id: string;
  quantity: number;
  merchandiseId: string;
  title: string;
  variantTitle: string;
  price: number;
  currencyCode: string;
  image: ShopifyProductImage | null;
}