# lib/api-zod/src/generated/types/shopifyCartLineInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ShopifyCartLineInput {
  merchandiseId: string;
  /**
     * @minimum 1
     * @maximum 99
     */
  quantity: number;
}