# lib/api-zod/src/generated/types/shopifyCartLinesInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ShopifyCartLineInput } from './shopifyCartLineInput';

export interface ShopifyCartLinesInput {
  /** @minItems 1 */
  lines: ShopifyCartLineInput[];
}