# lib/api-zod/src/generated/types/shopifyProduct.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ShopifyProductImage } from './shopifyProductImage';
import type { ShopifyProductVariant } from './shopifyProductVariant';

export interface ShopifyProduct {
  id: string;
  title: string;
  handle: string;
  description: string;
  image: ShopifyProductImage | null;
  price: number;
  currencyCode: string;
  variants: ShopifyProductVariant[];
}