# lib/api-zod/src/generated/types/shopifyProductImage.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ShopifyProductImage {
  url: string;
  /** @nullable */
  altText: string | null;
}