# lib/api-zod/src/generated/types/shopifyProductVariant.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ShopifyProductVariant {
  id: string;
  title: string;
  price: number;
  currencyCode: string;
  availableForSale: boolean;
}